+++ src/modules/ESP8266/ESP8266.h
@@ -51,7 +51,7 @@ class ESP8266: public TransportLayer {
     int16_t openTransportConnection(const char* host, const char* protocol, uint16_t port, uint16_t tcpKeepAlive = 0) override;
     int16_t closeTransportConnection() override;
     int16_t send(const char* data) override;
-    int16_t send(uint8_t* data, uint32_t len) override;
+    int16_t send(uint8_t* data, size_t len) override;
     size_t receive(uint8_t* data, size_t len, uint32_t timeout = 10000) override;
     size_t getNumBytes(uint32_t timeout = 10000, size_t minBytes = 10) override;
